scripts/generate/requirements.txt
diffusers>=0.27.0
torch>=2.2.0
accelerate>=0.28.0
Pillow>=10.0.0
